// src/app/model/plantao-model.ts
export class PlantaoModel {
  regiaoID: number = 0;
  regiao: string = '';
  dataInicial: string = '';
  dataFinal: string = '';
  magistrado: any;
  servidor: any;
  telefonePlantao: string = '';
  oficialJustica: any;
  cidade: string = '';
  emailPlantao: string = '';
  justificativa: string = '';
  responsavel: string = '';
  flag: string = '';
}